Correction Pass 3 complete + Crypto support + 4-gate chain tests
Part 1 fixes:
- ModelRegistry: wired settings.DUCKDB_PATH at all call sites (API + scheduler)
  Was pure in-memory dict — models vanished on restart
- ExperimentEngine: shared singleton via get_shared_engine() — routes and
  scheduler callbacks now share the same DB-backed instance
- 4th gate: SchedulerService.execute_champion_signal() — full 4-gate chain
  (config→RiskGovernor→KillSwitch→SCHEDULER_AUTO_EXECUTE) wired and tested
- verify_cleanup.sh: cancels open Alpaca orders + kills stray uvicorn
- Settings: extra='ignore' for not-yet-declared provider env keys

Crypto support (worker-data):
- AlpacaCryptoMarketDataProvider using CryptoHistoricalDataClient
- BTC/USD symbol format, synthetic fallback
- Fractional quantities: Order.qty/Fill.qty/Position.qty/Trade.size → float
- GTC time-in-force for crypto (detect by '/' in symbol)
- Broker get_positions: float(p.qty) — no truncation

4-gate chain tests (worker-tests):
- test_four_gate_chain.py: 5 tests proving each gate independently blocks
- test_order_service fractional qty: now accepted (float schema)
- 136/136 tests pass, ruff clean

Docs (worker-docs): STATUS/DECISIONS/BLOCKERS/CHANGELOG updated
  ADR-0015 (crypto provider), ADR-0016 (fractional qty)

+++ cudaquant/data/schemas.py
@@ -24,7 +24,7 @@ class Bar(BaseModel):
     high: float = Field(gt=0)
     low: float = Field(gt=0)
     close: float = Field(gt=0)
-    volume: int = Field(ge=0)
+    volume: float = Field(ge=0)
     frequency: BarFrequency
     vwap: float | None = None
     trade_count: int | None = None
@@ -36,7 +36,7 @@ class Trade(BaseModel):
     symbol: str
     timestamp: datetime
     price: float = Field(ge=0)
-    size: int = Field(ge=0)
+    size: float = Field(ge=0)
     exchange: str = ""
     conditions: list[str] | None = None
 
@@ -68,7 +68,7 @@ class Order(BaseModel):
     symbol: str
     side: OrderSide
     order_type: OrderType
-    qty: int = Field(gt=0)
+    qty: float = Field(gt=0)
     limit_price: float | None = Field(default=None, ge=0)
     time_in_force: str = "day"
 
@@ -79,7 +79,7 @@ class Fill(BaseModel):
     order_id: str
     symbol: str
     side: OrderSide
-    qty: int = Field(gt=0)
+    qty: float = Field(gt=0)
     price: float = Field(ge=0)
     timestamp: datetime
     commission: float = 0.0
@@ -89,7 +89,7 @@ class Position(BaseModel):
     """Current position in a symbol."""
 
     symbol: str
-    qty: int
+    qty: float
     avg_entry_price: float
     market_value: float | None = None
     unrealized_pnl: float | None = None
